requirements.txt
asgiref==3.5.1
books==0.1.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
Django==4.0.4
djangorestframework==3.13.1
idna==3.3
psycopg2==2.9.3
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
startapp==0.2.0.0
urllib3==1.26.9
